+++ packages/eslint-plugin-gestalt/src/__fixtures__/prefer-list/valid/valid.js
@@ -1,0 +1,5 @@
+import { List } from 'gestalt';
+
+export default function TestElement() {
+  return <List><List.Item text="Gestalt"/></List>;
+}